common: rework contract tests & mock
